refactor: clean up cross-link navigation UX and implementation
- Remove workspace→project crosslink (ambiguous with multiple projects per org)
- Move Org Settings crosslink into the tab list as a regular item, matching tab styling
- Drop sessionStorage hack and unnecessary useState/useEffect in SettingsLayout
- Delete duplicate useProject hook in features/settings/project/hooks; use canonical features/projects/hooks instead
