+++ frontend/ui/src/app/workspaces/[workspaceId]/settings/members/page.tsx
@@ -17,7 +17,6 @@ export default function WorkspaceSettingsMembersPage() {
         tabs={WORKSPACE_SETTINGS_TABS}
         activeTab="members"
         basePath={`/workspaces/${workspaceId}/settings`}
-        crossLink={{ label: "Project Settings", href: `/workspaces/${workspaceId}/projects` }}
       >
         <MembersTab workspaceId={workspaceId} />
       </SettingsLayout>
